requirements.txt
Flask>=2.2
python-docx>=0.8.11
rembg>=2.0.30
onnxruntime>=1.14.1
opencv-python>=4.7.0.72
pandas>=1.5.0
openpyxl>=3.1.0
